add / move some common functions/types
