+++ packages/common/src/types/index.ts
@@ -1,3 +1,5 @@
 export * from './old'
 export * from './api-routes'
 export * from './chatbot'
+
+export type Function<Props = void, Result = void> = (props: Props) => Result
